Safari-609.1.11.1/LayoutTests/fast/selectors/attribute-contains-value-matching-is-ascii-case-insensitive-expected.txt

When matching attributes case insensitively, it should be ASCII case insensitive. This test verifies the behavior when matching substring anywhere in the values (e.g. [a*="b"])

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing "[data-attribute*=Web-É-Kit]"
PASS document.querySelectorAll('#test-root [data-attribute*=Web-É-Kit]').length is 1
PASS document.querySelectorAll('#test-root [data-attribute*=Web-É-Kit]')[0].id is "target1"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[data-attribute*=b-É-K]"
PASS document.querySelectorAll('#test-root [data-attribute*=b-É-K]').length is 1
PASS document.querySelectorAll('#test-root [data-attribute*=b-É-K]')[0].id is "target1"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[data-attribute*=web-É-kit]"
PASS document.querySelectorAll('#test-root [data-attribute*=web-É-kit]').length is 1
PASS document.querySelectorAll('#test-root [data-attribute*=web-É-kit]')[0].id is "target2"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[data-attribute*=b-É-k]"
PASS document.querySelectorAll('#test-root [data-attribute*=b-É-k]').length is 1
PASS document.querySelectorAll('#test-root [data-attribute*=b-É-k]')[0].id is "target2"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[data-attribute*=WEB-É-KIT]"
PASS document.querySelectorAll('#test-root [data-attribute*=WEB-É-KIT]').length is 1
PASS document.querySelectorAll('#test-root [data-attribute*=WEB-É-KIT]')[0].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[data-attribute*=B-É-K]"
PASS document.querySelectorAll('#test-root [data-attribute*=B-É-K]').length is 1
PASS document.querySelectorAll('#test-root [data-attribute*=B-É-K]')[0].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[data-attribute*=Web-é-Kit]"
PASS document.querySelectorAll('#test-root [data-attribute*=Web-é-Kit]').length is 1
PASS document.querySelectorAll('#test-root [data-attribute*=Web-é-Kit]')[0].id is "target4"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[data-attribute*=b-é-K]"
PASS document.querySelectorAll('#test-root [data-attribute*=b-é-K]').length is 1
PASS document.querySelectorAll('#test-root [data-attribute*=b-é-K]')[0].id is "target4"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[data-attribute*=web-é-kit]"
PASS document.querySelectorAll('#test-root [data-attribute*=web-é-kit]').length is 1
PASS document.querySelectorAll('#test-root [data-attribute*=web-é-kit]')[0].id is "target5"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(10, 100, 200)"

Testing "[data-attribute*=b-é-k]"
PASS document.querySelectorAll('#test-root [data-attribute*=b-é-k]').length is 1
PASS document.querySelectorAll('#test-root [data-attribute*=b-é-k]')[0].id is "target5"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(10, 100, 200)"

Testing "[data-attribute*=Web-É-Kit i]"
PASS document.querySelectorAll('#test-root [data-attribute*=Web-É-Kit i]').length is 3
PASS document.querySelectorAll('#test-root [data-attribute*=Web-É-Kit i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [data-attribute*=Web-É-Kit i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [data-attribute*=Web-É-Kit i]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[data-attribute*=b-É-k i]"
PASS document.querySelectorAll('#test-root [data-attribute*=b-É-k i]').length is 3
PASS document.querySelectorAll('#test-root [data-attribute*=b-É-k i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [data-attribute*=b-É-k i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [data-attribute*=b-É-k i]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[data-attribute*=web-É-kit i]"
PASS document.querySelectorAll('#test-root [data-attribute*=web-É-kit i]').length is 3
PASS document.querySelectorAll('#test-root [data-attribute*=web-É-kit i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [data-attribute*=web-É-kit i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [data-attribute*=web-É-kit i]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[data-attribute*=b-É-k i]"
PASS document.querySelectorAll('#test-root [data-attribute*=b-É-k i]').length is 3
PASS document.querySelectorAll('#test-root [data-attribute*=b-É-k i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [data-attribute*=b-É-k i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [data-attribute*=b-É-k i]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[data-attribute*=WEB-É-KIT i]"
PASS document.querySelectorAll('#test-root [data-attribute*=WEB-É-KIT i]').length is 3
PASS document.querySelectorAll('#test-root [data-attribute*=WEB-É-KIT i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [data-attribute*=WEB-É-KIT i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [data-attribute*=WEB-É-KIT i]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[data-attribute*=B-É-K i]"
PASS document.querySelectorAll('#test-root [data-attribute*=B-É-K i]').length is 3
PASS document.querySelectorAll('#test-root [data-attribute*=B-É-K i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [data-attribute*=B-É-K i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [data-attribute*=B-É-K i]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[data-attribute*=Web-é-Kit i]"
PASS document.querySelectorAll('#test-root [data-attribute*=Web-é-Kit i]').length is 2
PASS document.querySelectorAll('#test-root [data-attribute*=Web-é-Kit i]')[0].id is "target4"
PASS document.querySelectorAll('#test-root [data-attribute*=Web-é-Kit i]')[1].id is "target5"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(10, 100, 200)"

Testing "[data-attribute*=b-é-k i]"
PASS document.querySelectorAll('#test-root [data-attribute*=b-é-k i]').length is 2
PASS document.querySelectorAll('#test-root [data-attribute*=b-é-k i]')[0].id is "target4"
PASS document.querySelectorAll('#test-root [data-attribute*=b-é-k i]')[1].id is "target5"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(10, 100, 200)"

Testing "[data-attribute*=web-é-kit i]"
PASS document.querySelectorAll('#test-root [data-attribute*=web-é-kit i]').length is 2
PASS document.querySelectorAll('#test-root [data-attribute*=web-é-kit i]')[0].id is "target4"
PASS document.querySelectorAll('#test-root [data-attribute*=web-é-kit i]')[1].id is "target5"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(10, 100, 200)"

Testing "[data-attribute*=b-é-k i]"
PASS document.querySelectorAll('#test-root [data-attribute*=b-é-k i]').length is 2
PASS document.querySelectorAll('#test-root [data-attribute*=b-é-k i]')[0].id is "target4"
PASS document.querySelectorAll('#test-root [data-attribute*=b-é-k i]')[1].id is "target5"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(10, 100, 200)"

Testing "[multiple*=WEB-é-KIT]"
PASS document.querySelectorAll('#test-root [multiple*=WEB-é-KIT]').length is 3
PASS document.querySelectorAll('#test-root [multiple*=WEB-é-KIT]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple*=WEB-é-KIT]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple*=WEB-é-KIT]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[multiple*=B-é-K]"
PASS document.querySelectorAll('#test-root [multiple*=B-é-K]').length is 3
PASS document.querySelectorAll('#test-root [multiple*=B-é-K]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple*=B-é-K]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple*=B-é-K]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[multiple*=Web-é-Kit]"
PASS document.querySelectorAll('#test-root [multiple*=Web-é-Kit]').length is 3
PASS document.querySelectorAll('#test-root [multiple*=Web-é-Kit]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple*=Web-é-Kit]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple*=Web-é-Kit]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[multiple*=b-é-k]"
PASS document.querySelectorAll('#test-root [multiple*=b-é-k]').length is 3
PASS document.querySelectorAll('#test-root [multiple*=b-é-k]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple*=b-é-k]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple*=b-é-k]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[multiple*=Web-é-kit]"
PASS document.querySelectorAll('#test-root [multiple*=Web-é-kit]').length is 3
PASS document.querySelectorAll('#test-root [multiple*=Web-é-kit]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple*=Web-é-kit]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple*=Web-é-kit]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[multiple*=b-é-K]"
PASS document.querySelectorAll('#test-root [multiple*=b-é-K]').length is 3
PASS document.querySelectorAll('#test-root [multiple*=b-é-K]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple*=b-é-K]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple*=b-é-K]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[multiple*=web-É-kit]"
PASS document.querySelectorAll('#test-root [multiple*=web-É-kit]').length is 2
PASS document.querySelectorAll('#test-root [multiple*=web-É-kit]')[0].id is "target4"
PASS document.querySelectorAll('#test-root [multiple*=web-É-kit]')[1].id is "target5"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(10, 100, 200)"

Testing "[multiple*=b-É-K]"
PASS document.querySelectorAll('#test-root [multiple*=b-É-K]').length is 2
PASS document.querySelectorAll('#test-root [multiple*=b-É-K]')[0].id is "target4"
PASS document.querySelectorAll('#test-root [multiple*=b-É-K]')[1].id is "target5"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(10, 100, 200)"

Testing "[multiple*=web-É-Kit]"
PASS document.querySelectorAll('#test-root [multiple*=web-É-Kit]').length is 2
PASS document.querySelectorAll('#test-root [multiple*=web-É-Kit]')[0].id is "target4"
PASS document.querySelectorAll('#test-root [multiple*=web-É-Kit]')[1].id is "target5"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(10, 100, 200)"

Testing "[multiple*=b-É-k]"
PASS document.querySelectorAll('#test-root [multiple*=b-É-k]').length is 2
PASS document.querySelectorAll('#test-root [multiple*=b-É-k]')[0].id is "target4"
PASS document.querySelectorAll('#test-root [multiple*=b-É-k]')[1].id is "target5"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(10, 100, 200)"

Testing "[multiple*=WEB-é-KIT i]"
PASS document.querySelectorAll('#test-root [multiple*=WEB-é-KIT i]').length is 3
PASS document.querySelectorAll('#test-root [multiple*=WEB-é-KIT i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple*=WEB-é-KIT i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple*=WEB-é-KIT i]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[multiple*=B-é-k i]"
PASS document.querySelectorAll('#test-root [multiple*=B-é-k i]').length is 3
PASS document.querySelectorAll('#test-root [multiple*=B-é-k i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple*=B-é-k i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple*=B-é-k i]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[multiple*=Web-é-Kit i]"
PASS document.querySelectorAll('#test-root [multiple*=Web-é-Kit i]').length is 3
PASS document.querySelectorAll('#test-root [multiple*=Web-é-Kit i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple*=Web-é-Kit i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple*=Web-é-Kit i]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[multiple*=b-é-K i]"
PASS document.querySelectorAll('#test-root [multiple*=b-é-K i]').length is 3
PASS document.querySelectorAll('#test-root [multiple*=b-é-K i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple*=b-é-K i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple*=b-é-K i]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[multiple*=Web-é-kit i]"
PASS document.querySelectorAll('#test-root [multiple*=Web-é-kit i]').length is 3
PASS document.querySelectorAll('#test-root [multiple*=Web-é-kit i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple*=Web-é-kit i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple*=Web-é-kit i]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[multiple*=B-é-K i]"
PASS document.querySelectorAll('#test-root [multiple*=B-é-K i]').length is 3
PASS document.querySelectorAll('#test-root [multiple*=B-é-K i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple*=B-é-K i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple*=B-é-K i]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[multiple*=web-É-kit i]"
PASS document.querySelectorAll('#test-root [multiple*=web-É-kit i]').length is 2
PASS document.querySelectorAll('#test-root [multiple*=web-É-kit i]')[0].id is "target4"
PASS document.querySelectorAll('#test-root [multiple*=web-É-kit i]')[1].id is "target5"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(10, 100, 200)"

Testing "[multiple*=b-É-K i]"
PASS document.querySelectorAll('#test-root [multiple*=b-É-K i]').length is 2
PASS document.querySelectorAll('#test-root [multiple*=b-É-K i]')[0].id is "target4"
PASS document.querySelectorAll('#test-root [multiple*=b-É-K i]')[1].id is "target5"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(10, 100, 200)"

Testing "[multiple*=web-É-Kit i]"
PASS document.querySelectorAll('#test-root [multiple*=web-É-Kit i]').length is 2
PASS document.querySelectorAll('#test-root [multiple*=web-É-Kit i]')[0].id is "target4"
PASS document.querySelectorAll('#test-root [multiple*=web-É-Kit i]')[1].id is "target5"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(10, 100, 200)"

Testing "[multiple*=b-É-k i]"
PASS document.querySelectorAll('#test-root [multiple*=b-É-k i]').length is 2
PASS document.querySelectorAll('#test-root [multiple*=b-É-k i]')[0].id is "target4"
PASS document.querySelectorAll('#test-root [multiple*=b-É-k i]')[1].id is "target5"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(10, 100, 200)"

PASS successfullyParsed is true

TEST COMPLETE
